feat: enhance user profile management and navigation with new fields …

+++ src/components/shared/image-uploader.tsx
@@ -66,9 +66,11 @@ export const ImageUploader = ({
 
   const onDelete = (url: string) => {
     onRemove(url);
-    deleteObject(ref(storage, url)).then(() => {
-      toast.success("Image deleted successfully");
-    });
+    if (url.startsWith('https://firebasestorage.googleapis.com/')) {
+      deleteObject(ref(storage, url)).then(() => {
+        toast.success("Image deleted successfully");
+      });
+    }
   };
 
   return (
